Setup yaql and jinja json functions for orquesta
Wire the json related functions in orquesta and include unit and integration tests to ensure they are configured correctly. Consolidate the json related functions and place them in the same module with the data functions.
