+++ contrib/runners/orquesta_runner/tests/unit/test_functions.py
@@ -122,12 +122,13 @@ class OrquestaFunctionTest(st2tests.DbTestCase):
 
         # Check workflow output.
         expected_output = {
-            'data_json_str_1': '{"foo": "bar"}',
-            'data_json_str_2': '{"foo": "bar"}',
-            'data_json_obj_1': {'foo': 'bar'},
-            'data_json_obj_2': {'foo': 'bar'},
-            'data_yaml_str_1': 'foo: bar\n',
-            'data_pipe_str_1': '{"foo": "bar"}'
+            'data_json_str_1': '{"foo": {"bar": "foobar"}}',
+            'data_json_str_2': '{"foo": {"bar": "foobar"}}',
+            'data_json_obj_1': {'foo': {'bar': 'foobar'}},
+            'data_json_obj_2': {'foo': {'bar': 'foobar'}},
+            'data_yaml_str_1': 'foo:\n  bar: foobar\n',
+            'data_query_1': ['foobar'],
+            'data_pipe_str_1': '{"foo": {"bar": "foobar"}}'
         }
 
         self.assertDictEqual(wf_ex_db.output, expected_output)
@@ -170,11 +171,12 @@ class OrquestaFunctionTest(st2tests.DbTestCase):
 
         # Check workflow output.
         expected_output = {
-            'data_json_str_1': '{"foo": "bar"}',
-            'data_json_str_2': '{"foo": "bar"}',
-            'data_json_obj_1': {'foo': 'bar'},
-            'data_json_obj_2': {'foo': 'bar'},
-            'data_yaml_str_1': 'foo: bar\n'
+            'data_json_str_1': '{"foo": {"bar": "foobar"}}',
+            'data_json_str_2': '{"foo": {"bar": "foobar"}}',
+            'data_json_obj_1': {'foo': {'bar': 'foobar'}},
+            'data_json_obj_2': {'foo': {'bar': 'foobar'}},
+            'data_yaml_str_1': 'foo:\n  bar: foobar\n',
+            'data_query_1': ['foobar']
         }
 
         self.assertDictEqual(wf_ex_db.output, expected_output)
