+++ st2common/st2common/expressions/functions/data.py
@@ -16,13 +16,18 @@
 from __future__ import absolute_import
 
 import json
+import jsonpath_rw
 import mongoengine
 import six
 import yaml
 
+
 __all__ = [
     'from_json_string',
     'from_yaml_string',
+    'json_escape',
+    'jsonpath_query',
+    'to_complex',
     'to_json_string',
     'to_yaml_string',
 ]
@@ -64,3 +69,39 @@ def to_yaml_string(value, indent=None, allow_unicode=True):
         options['allow_unicode'] = allow_unicode
 
     return yaml.safe_dump(value, **options)
+
+
+def json_escape(value):
+    """ Adds escape sequences to problematic characters in the string
+    This filter simply passes the value to json.dumps
+    as a convenient way of escaping characters in it
+    However, before returning, we want to strip the double
+    quotes at the ends of the string, since we're not looking
+    for a valid JSON value at the end, just conveniently using
+    this function to do the escaping. The value could be any
+    arbitrary value
+    """
+
+    return json.dumps(value).strip('"')
+
+
+def jsonpath_query(value, query):
+    """Extracts data from an object `value` using a JSONPath `query`.
+    :link: https://github.com/kennknowles/python-jsonpath-rw
+    :param value: a object (dict, array, etc) to query
+    :param query: a JSONPath query expression (string)
+    :returns: the result of the query executed on the value
+    :rtype: dict, array, int, string, bool
+    """
+    expr = jsonpath_rw.parse(query)
+
+    matches = [match.value for match in expr.find(value)]
+
+    if not matches:
+        return None
+
+    return matches
+
+
+def to_complex(value):
+    return json.dumps(value)
