+++ st2tests/st2tests/fixtures/packs/orquesta_tests/actions/workflows/yaql-data-functions.yaml
@@ -3,7 +3,7 @@ version: 1.0
 description: A basic workflow testing data related functions.
 
 vars:
-  - data_json_str_1: '{"foo": "bar"}'
+  - data_json_str_1: '{"foo": {"bar": "foobar"}}'
 
 tasks:
   task1:
@@ -15,3 +15,4 @@ output:
   - data_yaml_str_1: <% to_yaml_string(ctx(data_json_obj_1)) %>
   - data_json_obj_2: <% from_yaml_string(ctx(data_yaml_str_1)) %>
   - data_json_str_2: <% to_json_string(ctx(data_json_obj_2)) %>
+  - data_query_1: <% jsonpath_query(ctx(data_json_obj_2), '$.foo.bar') %>
